requirements.txt
python-dotenv>=1.0.1
yfinance>=0.2.54
pandas>=2.2.3
requests>=2.32.3
openai>=2.0.0
openpyxl>=3.1.5
